+++ src/routes/work.ts
@@ -1,5 +1,5 @@
 import { Router } from "express";
-import { add, all } from "../controllers/tasks";
+import { add, one, all } from "../controllers/work";
 import { isAdmin } from "../middlewares/auth";
 
 
@@ -7,5 +7,6 @@ const router = Router();
 
 router.post("/", isAdmin, add);
 router.get("/", all);
+router.get("/:id", one);
 
 export default router;
